requirements.txt
fastapi
uvicorn[standard]
python-multipart
inference-sdk
pillow
python-dotenv
